Ticket #47960 - cookie_change_info returns random negative number if there was no change in a tree
Description: An additional fix for the type mismatch for the change
numbers.  Change Number is declared as "unsigned long" in the Retro
Changelog plugin, while cookie_change_info in the Content Sync plugin
is "int", which could end up with a negative number when the change
number passes (2^31 - 1).

Changing the type of cookie_change_info to "unsigned long".

https://fedorahosted.org/389/ticket/47960

Reviewed by rmeggins@redhat.com and tbordaz@redhat.com (Thank you,
Rich and Thierry!!)