Ticket #47905 - Bad manipulation of passwordhistory
Description: The patch was provided by German Parente (gparente@redhat.com)
https://fedorahosted.org/389/ticket/47905#comment:6

If a value of password policy attribute (e.g., passwordhistory) is
accidentally deleted, it causes a null reference in pw_val2scheme.
This patch checks whether val is NULL or not.  If the given val is
NULL, pw_val2scheme returns NULL.

https://fedorahosted.org/389/ticket/47905

Reviewed by nhosoi@redhat.com.