Ticket #47934 - nsslapd-db-locks modify not taking into account.
Description: Updating the value of nsslapd-db-locks was not detected
at the server's startup time.  As suggested in
    https://fedorahosted.org/389/ticket/47934#comment:1
the value is stored in the guardian file as "locks:20000" at the shut-
down time.  When restarting the server, the nsslapd-db-locks value in
cn=config,cn=ldbm database,cn=plugins,cn=config is compared with the
stored value (20000, in this example) and if they don't match, the
lock table is recreated.  The change is logged in the error log:
    [...] - resizing max db lock count: 20000 -> 30000

https://fedorahosted.org/389/ticket/47934

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)

(cherry picked from commit f4437defacf8812c5f95869147a99f8412dc4a07)