Ticket #47836 - Do not return '0' as empty fallback value of nsds5replicalastupdatestart and nsds5replicalastupdatestart
Bug Description: In a replica agreement entry, if any of the
nsds5replicaLastUpdateStart, nsds5replicaLastUpdateEnd,
nsds5replicaLastInitStart or nsds5replicaLastInitEnd are not set,
it's initialized as "0" which is invalid.  It should be in the
LDAP Generalized time format.

Fix Description: When these attributes are not set, instead of "0"
they are initialized with the epoch time 0 == 19700101000000Z.

https://fedorahosted.org/389/ticket/47836

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)