Ticket #47957 - Make ReplicaWaitForAsyncResults configurable
Description: Introducing a config attr nsDS5ReplicaWaitForAsyncResults
to the agreement entry.
  dn: cn=<AGREEMENT>,cn=replica,cn="<SUFFIX>",cn=mapping tree,cn=config
  nsDS5ReplicaWaitForAsyncResults: <integer in millisecond>

Prior to this patch, supplier sleeps 1 second if it finds the response
from consumer is not ready.  1 second could be too long if higher
replication throughput is required.

This patch makes the waiting time configurable, and change the default
to 100 millisecond.  If the attribute nsDS5ReplicaWaitForAsyncResults
does not exist or the value is 0, the default value is set.

https://fedorahosted.org/389/ticket/47957

Reviewed by rmeggins@redhat.com (Thank you!!)