Ticket #48133 - Non tombstone entry which dn starting with "nsuniqueid=...," cannot be deleted
Bug Description: Trying to delete an entry which DN starts with
"nsuniqueid=...," but no objectclass=nsTombstone fails with
"Turning a tombstone into a tombstone!", which is indeed not.

Fix Description: This patch checks the entry and if it does not
have "objectclass=nsTombstone", the entry is not treated as a
tombstone.  Also, if the DN already has the entry's nsuniqueid
at the beginning, it does not get appended to avoid the duplicate.

Note: Adding an entry which DN starts with "nsuniqueid" and no
nsTombstone objectclass fails since such an entry is rejected in
check_rdn_for_created_attrs called from do_add.

https://fedorahosted.org/389/ticket/48133

Reviewed by rmeggins@redhat.com (Thank you, Rich!!)

(cherry picked from commit f14fb8b6f71d2a87b650064f15099a2cc2dfd38a)