Ticket #47928 - Disable SSL v3, by default [389-ds-base-1.2.11 only]
Description: commit 17fc03cf1101135b99234f17efd3eb746626be1a introduced a
defect "bz 1210996 - TLS1 can't be turned off".  This patch correctly sets
SSLv3 to minimum and maximum version if nsTLS1 is off.  If both nsSSL3 and
nsTLS1 is off, it forces to enable nsTLS1 and log it in the error log.