Ticket #48109 - substring index with nssubstrbegin: 1 is not being used with filters like (attr=x*)
Description: In the commit ae69e4aa1dcec1c5f79af502b77d9e19a3d348ee,
when pblock is passed to slapi_attr_assertion2keys_sub_sv_pb from the
caller, the original slapdplugin is overwritten in the function.  In
the case, the original slapdplugin has to be backed up and restored.

https://fedorahosted.org/389/ticket/48109

Reviewed by rmeggins@redhat.com (Thank you, Rich!!)