Ticket #48190 - idm/ipa 389-ds-base entry cache converges to 500 KB in dblayer_is_cachesize_sane
Description: This issue was introduced by the fix for Ticket 47499
commit 1e035d1111f6abcb87e760a2b9e41fa9e05a7ebd.

The function dblayer_is_cachesize_sane was originally implemented for
db cache to check if the given db cache size is larger than the available
memory or not.  The function resets the size to the available memory size
if it is larger.  Also, considering the extra metadata size needed for the
db cache, it multiplies by 0.8 every time it starts the server.  It is not
needed even for the db cache.  The code is old and we don't have to save
the memory there.  Thus, this patch removes the resetting code.

https://fedorahosted.org/389/ticket/48190

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)