Ticket #48191 - RFE: Adding nsslapd-maxsimplepaged-per-conn
Description: Asynchronous simple paged results requests could add too much load
the server can handle. Adding a config parameter to restrict the requests.

  cn=config
  nsslapd-maxsimplepaged-per-conn: INT

If nsslapd-maxsimplepaged-per-conn is configured with a positive integer,
Asynchronous simple paged results requests per connection is limitted by the
value.  If the requests exceed the value, it returns LDAP_UNWILLING_TO_PERFORM.

If the value is negative, there is no limit (default behaviour).
If the value is 0, a simple paged results is disabled.

https://fedorahosted.org/389/ticket/48191

Reviewed by rmeggins@redhat.com (Thank you, Rich!)