Ticket #48212 - Dynamic nsMatchingRule changes had no effect on the attrinfo thus following reindexing, as well.
Description: When nsMatchingRule was dynamically updated in an index entry,
the value was set to the configuration but was not applied to the attribute
info.  On-line reindexing following the nsMatchingRule change actually ignored
the setting. On the other hand, the standalone utility dbverify independently
picked up the nsMatchingRule from the configuration and generated the attribute
info, which expected the index reindexed based upon the new nsMatchingRule. But
it was actually not and dbverify reported the index corruption.

This patch applies the changes to the attribute info when nsMatchingRule is
modified.

https://fedorahosted.org/389/ticket/48212

Reviewed by rmeggins@redhat.com (Thank you, Rich!)

(cherry picked from commit d9679725e69df1d191864ca00bad6b79b13e7362)
(cherry picked from commit d15beff66fd28902bd8ca80af12ad76a7ecbe57d)