Ticket #48216 - crash in ns-slapd when deleting winSyncSubtreePair from sync agreement
Description: In free_subtree_pairs, the condition for stopping the
loop to clean up the AD and DS subtree dn was incomplete.

This patch checks the AD and DS subtree dn and if any of the pair
is NULL, it stops the clean up.  Related to the issue, more checks
for the validation of the winSyncSubtreePair is added so that any
single valued cases are ignored with an error log.
[single valued case examples]
  winSyncSubtreePair: ou=People,dc=anytree
  winSyncSubtreePair: ou=People,dc=anytree:
  winSyncSubtreePair: :ou=People,dc=anytree

https://fedorahosted.org/389/ticket/48216

Reviewed by rmeggins@redht.com (Thank you, Rich!!)

(cherry picked from commit 6dce81e77a47dc23e0b825952c97112039f45201)
(cherry picked from commit 6d177bf359c022f1e46d575c6fe3ad3d97f1cfeb)