Ticket #48223 - Winsync fails when AD users have multiple spaces (two)inside the value of the rdn attribute
Description: When the dirsync search returns a remote entry, winsync
search the entry with DN to retrieve the whole attribute value pairs.
The DN used for the search was normalized which replaced multiple white-
spaces with one in the DN.  This patch does not used the normalized DN,
but the same DN given by AD.

The DN normalization behaviour was introduced to fix a ticket #529 -
dn normalization must handle multiple space characters in attributes.

Added additional debugging to get the info which entry failed to sync.

https://fedorahosted.org/389/ticket/48223

Reviewed by rmeggins@redhat.com (Thank you, Rich!!)