Ticket #48228 - wrong password check if passwordInHistory is decreased.
Bug Description: When N passwords to be remembered (passwordInHistroy)
and N passwords are remembered, decreasing the passwordInHistory value
to M (< N) does not allow to use the oldest password which should have
been discarded from the history and should be allowed.

Fix Description: Before checking if the password is in the history or
not, adding a check the passwordInHistory value (M) is less than the
count of passwords remembered (N).  If M < N, discard the (N-M) oldest
passwords.

https://fedorahosted.org/389/ticket/48228

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)