Ticket #47511 - bashisms in 389-ds-base admin scripts
Description by mvocu (Thank you):
The shell scripts in 389-ds-base/ldap/admin/src/scripts use 'source'
to source common scripts; the 'source' keyword is bash-specific (or
c-shell, if memory serves). The interpreter is set to /bin/sh, which
is not guaranteed to be bash (and at least on Debian 7.1 it is dash).
The 'source' keyword can be replaced by '.', which should work.

The patch was provided by tjaalton@ubuntu.com (Thank you, Timo!).

https://fedorahosted.org/389/ticket/47511

Reviewed and tested by nhosoi@redhat.com.