Ticket #47569 - Added a testcase to ACL testsuite
Description: The attribute defined in the targetattr keyword of an ACI
is checked against the schema to make sure it is a defined attribute
when you are adding a new ACI.  If you want to use an attribute subtype,
the ACI is rejected since the attribute with subtype is not defined in
the schema.  We should strip off the subtype when we validate the
targetattr keyword against the schema.

Test description:
1. Define two attributes in the schema
    - first will be a targetattr
    - second will be a userattr
2. Add an ACI with an attribute subtype
    - or language subtype
    - or binary subtype
    - or pronunciation subtype

Signed-off-by: Mark Reynolds <mreynolds@redhat.com>
(cherry picked from commit 0c4eafbc945ae4252886ba8546665a79206f3f83)