Ticket #48252 - db2index creates index entry from deleted records
Bug Description: When an entry is deleted, its indexed attribute values
are also removed from each index file.  But if the entry is turned to be
a tombstone entry, reindexing adds the removed attribute value back to
the index file.

Fix Description: In the reindexing function ldbm_back_ldbm2index, if the
entry is a tombstone entry, it skips reindexing operation unless the target
attribute is entryrdn or objectclass.  If it is reindexing the objectclass,
it indexes just for the key "=nstombstone".

https://fedorahosted.org/389/ticket/48252

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)