Include necessary autogenerated files
Some of the previous patches assumed non committed to git files and the use
of autoreconf -fi to build so some files were not added (and indeed actively
ignored in .gitignore. This caused make dist and make rpms to fail from a
git checkout where autoreconf -fi is not run.

This patch adds the needed macros to successfully run make dist and make rpms
It also removes files from .gitignore tht are committed to git as git status
always checks for modifications to committed files anyway.

Signed-off-by: Simo Sorce <simo@redhat.com>