Ticket #48192 - Individual abandoned simple paged results request has no chance to be cleaned up
Description: If CONN_FLAG_PAGEDRESULTS_ABANDONED is set to pr_flags,
the search results in the pagedresults handle is supposed to have been
cleaned up.  But when there is a contention, there is a case that it
is reset with the already released search results.  This patch adds an
additional check for abandoned flag in pagedresults_set_search_result.
If the pagedresults handle shows it is abandoned, the search results
is not set to the handle unless it is for cleaning up with NULL.

https://fedorahosted.org/389/ticket/48192

Reviewed by rmeggins@redhat.com (Thanks, Rich!!)

(cherry picked from commit 6e453918e82af6c597390aebf92a8eb3283c3591)
(cherry picked from commit 96b9b6794e0a6bfa0d74c84f6c80131c4f820fa7)