Ticket #48304 - ns-slapd - LOGINFO:Unable to remove file
Description: In the log rotation, if a log file to be deleted does
not exist, the "Unable to remove file" is logged in the error log,
which is not necessary.

This patch updates the logging code to suppress the unnecessary log
messages as well as replace with more detailed ones.

https://fedorahosted.org/389/ticket/48304

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)