Ticket #48287 - Double free while adding entries (1.2.11 only)
Description: If a callback at SLAPI_PLUGIN_BE_TXN_*_ADD_FN fails and the
adding-entry is in a cache, the ldbm_back_add is supposed to remove the
adding-entry from the cache and free it.  The issue was fixed in 1.3.1
and newer by these tickets:
 Ticket #47808 - If be_txn plugin fails in ldbm_back_add, adding entry is double freed.
 Ticket #47815 - Add operations rejected by betxn plugins remain in cache
which were not backported to 1.2.11.

https://fedorahosted.org/389/ticket/48287

Reviewed by tbordaz@redhat.com (Thank you, Thierry!!)

(cherry picked from commit 265c6e399016ad4a46c8709d32367b9c30ea57cf)