Ticket #48285 - The dirsrv user/group should be created in rpm %pre, and ideally with fixed uid/gid
Description: Adding user dirsrv (uid 389) and group dirsrv (gid 389)
if they do not exist.  If the uid and the gid are in use, increase
one by one and find unused id for them.  If they exists, no change
is made.  The default user and group pair (nobody, nobody) is also
replaced with (dirsrv, dirsrv).

https://fedorahosted.org/389/ticket/48285

Reviewed by wibrown@redhat.com (Thank you, William!)