Ticket 48321 - Improve is_a_dn check to prevent mistakes with lib389 auth https://fedorahosted.org/389/ticket/48321
Bug Description:   is_a_dn previously had a weak check to determine if a binddn
was in fact, a dn. This improves the check's quality to be far more thorough,
which is needed for the rest389 api.

Fix Description:  We now use the ldap.dn function to check the dn is valid, as
well as potentially allowing anonymous ("") as a valid dn for binding.

Author: wibrown

Review by: ???