Ticket 47840 - add configure option to disable instance specific scripts https://fedorahosted.org/389/ticket/47840
Bug Description:  Now that ​https://fedorahosted.org/389/ticket/528 is
fixed, the next step is to allow building the server with the instance specific
scripts disabled.

Fix Description:  Instead of defining a configure option, we provide a new
option in setup-ds.pl, slapd.InstScriptsEnabled, which defaults to false. All
new installs of 389 will NOT install with a inst_dir nor the scripts that are in
that directory.

Additionally, this change fixes setup-ds.pl to correctly use the sbindir scripts
to start/stop the server instance during installation.

Finally, we add support for setup-ds.pl so that in --update if the inst_dir
exists, scripts will be updated, but if it does not exist, no action is taken.

In time, we will alter --update to *remove* the scripts within inst_dir during
the update (We have no way of knowing if a customer has put custom scripts in
inst_dir)

Example:
/opt/dirsrv/sbin/setup-ds.pl slapd.InstScriptsEnabled=false

Author: wibrown

Review by: nhosoi (Thanks!)