Ticket #48328 - RFE: Allow RHDS to be setup using a DNS CNAME alias for General.FullMachineName
Description: When the Directory Server is installed, its FQDN may need to
be an alias (CNAME) to let it run behind the proxy such as a load balancer.
The current setup script does not allow it.  This patch addes to check the
given FQDN with CNAME in addition to doing the reverse lookup.

https://fedorahosted.org/389/ticket/48328

Reviewed by wibrown@redhat.com (Thank you, William!!)