Ticket #48317: SELinux port labeling retry attempts are excessive
https://fedorahosted.org/389/ticket/48317

Bug Description:  In dscreate.pm we attempt to label the ldap_port_t type 60
times in the case of a failure. This is excessive, and it means the setup-ds.pl
appears to hang in certain cases.

Fix Description:
Reduce this number to 5 attempts, and when debug is enabled, display the amount
of attempts remaining.

Author: wibrown@redhat.com

Review by: nhosoi@redhat.com (Thank you Noriko!)