Ticket 48145 - Allow merged logging of audit events
Bug Description:  The auditfail logging should be able to be directed to the
same audit file, or to it's own seperate file.

Fix Description:  When nsslapd-auditfaillog is not specified the value of
nsslapd-auditlog will be used for audit and auditfail events. If auditfaillog
is specified, all results with RC != LDAP_SUCCESS (0) will go to the auditfail
handler.

https://fedorahosted.org/389/ticket/48145

Author: wibrown

Review by: mreynolds (Thanks!)