Ticket 48332 - allow users to specify to relax the FQDN constraint
Bug Description:  There are situations when the machine name for ds may not
match the dns name. In these cases we should allow installation without the
strict hostname checks we carry out.

Fix Description:  Add a new option, General.StrictHostCheck which defaults
to true. If true, host name checking is carried out. If false, it is disabled
and any hostname in General.FullMachineName is considered valid.

https://fedorahosted.org/389/ticket/48332

Author: wibrown

Review by: rmeggins (Thanks!)