Ticket 48364 - Fix test failures
Description: New failures appear in the test suites. They should be
fixed as soon as possible. More new features are coming to lib389(like
REST or krb5 support) and they require a proper CI testing.

Fix description: Remove "+" sign from multiline string in the
lib389/backend.py, because it makes an impossible to replace string with
variables on the second line.

- tests/backend_test.py
Fix exception naming in the tests, because it was updated with recent
patches within lib389/backend.py module.

- tests/suffix_test.py
Refactor code to the pytest compatibility.

https://fedorahosted.org/389/ticket/48364

Reviewed by: wibrown (Thanks!)