Ticket 48333 - Simplify accessiblitiy of tests for ds
Bug Description:  Previously, it was hard to run the full test suite without
scripting or other external means. We want to be able to run

py.test dirsrvtests

To test the full scope of the application. We can also use this to trigger
specific tests:

py.test dirsrvtests -k test_ticket48109_1

Fix Description:  To achieve this, we create __init__.py in all the needed
folders, to make the python path accessible by py.test to discover all the
tests. Additionally, we add __pycache__ and *.pyc to gitignore

https://fedorahosted.org/389/ticket/48333

Author: wibrown

Review by: mreynolds (Thanks!)