Ticket 48280 - enable logging of internal ops in the audit log
Bug Description:  many plugins add and modify entries, but these changes are not
visible in the audit log. It can be very useful to have these changes available.

Fix Description:  This feature actually already existed:

nsslapd-logAccess
nsslapd-logAudit

On a plugin, will turn on the respective logs.

This change brings the audit log in line with access to respect the value of

nsslapd-plugin-logging

Such that when plugin-logging is enabled globally in cn=config, plugins will
log based on the values in:

nsslapd-accesslog-logging-enabled
nsslapd-auditlog-logging-enabled
nsslapd-auditfaillog-logging-enabled

https://fedorahosted.org/389/ticket/48280

Author: wibrown

Review by: ???