Ticket #142 - [RFE] Default password syntax settings don't work with fine-grained policies
Description: When a fine-grained password syntax is not set, there is no
syntax restriction for the passwords to be added or modified even though
the global password syntax is configured.

This patch introducing a config parameter:
  cn=config.
  nsslapd-pwpolicy-inherit-global: on | off
If on, the fine-grained password syntax inherits the global password
syntax if the global one is configured.
If off, the inheritance does not occur.  This is the current behaviour.
By default, it is off.

https://fedorahosted.org/389/ticket/142

Reviewed by firstyear@redhat.com (Thank you, William!!)