Ticket #48294 - Linked Attributes plug-in - won't update links after MODRDN operation
Description: When an entry is renamed and the DN is a value of linktype,
since the linktype value is under control by the user, the value needs to
be manually modified to adjust to the new DN.

But the modification failed in linked_attrs_mod_backpointers due to the
too tight error checking.  This patch allows LDAP_NO_SUCH_OBJECT for the
deletion of the old DN and LDAP_TYPE_OR_VALUE_EXISTS for adding a managed
entry which exists prior to the modification.

https://fedorahosted.org/389/ticket/48294

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)