Ticket #48295 - Entry cache is not rolled back -- Linked Attributes plug-in - wrong behaviour when adding valid and broken links
Bug Description: When multiple link attribute values are added or deleted
and if some of the op was successfully and one fails, the succeeded operation's
managed type appears in the managed entry although all of the link attribute
operation were rolled back.  The managed type indeed does not exist in the
managed entry.  That just in the entry cache.  Once the server is restarted
the changes disappear.

Fix Description: If the first link attribute value operation fails, it stops
the operation and returns the failure.  If the second or the later operation
fails, the succeeded operation up to the previous one are reverted.

https://fedorahosted.org/389/ticket/48295

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)