Ticket 48390 - RFE Improvements to lib389 monitor features for rest389
Bug Description:  We need to improve the monitor features of lib389, so that we
can expose more data for the use of the new admin system and cli tools. This
change will allow this to occur. Additionally, we can now base other advisory
tools on this for tuning etc.

Fix Description:  Implement more monitoring features.

https://fedorahosted.org/389/ticket/48390

Author: wibrown

Acknowledgements: I would like to thank the following staff of the University
of Adelaide for their advice on monitoring and needs that customers have with
ds systems:

* Mark Larsen
* David Monro
* Tim Bowen
* Matt Salkeld

Review by: ???