Ticket 48395 - ASAN - Use after free in uiduniq 7bit.c
Bug Description:   Asan detected a use after free in 7bit.c during a modrdn
operation. This may cause the directory to crash if a specially crafted modrdn
request is made, or may be exploitable in some other way.

Fix Description:  We move the issue_error directive before the slapi_entry is
freed, which means that we are using before the free instead.

https://fedorahosted.org/389/ticket/48395

Author: wibrown

Review by: nhosoi (Thank you!)