Ticket 48397 - ds make rpms doesn't work in a prefixed build env
Bug Description:  if you run a prefixed build, make rpms does not work.

Additionally, rpmbuild throws invalid %if errors with this .spec

Fix Description:  This fixes the build paths to work regardless of prefix build
or not. This also fixes the rpm if macros to work, and the missing dirsrvtests
from the dist component of Makefile.am

https://fedorahosted.org/389/ticket/48397

Author: wibrown

Review by: mreynolds